Can not delete event once attendee is invited or paid (#166)

+++ server/services/event.py
@@ -128,9 +128,16 @@ class EventService:
         if not db_event:
             raise NotFoundError("Event not found.")
 
-        try:
-            attendees = self.attendee_service.get_attendees_for_event(db_event.id)
+        attendees = self.attendee_service.get_attendees_for_event(db_event.id)
 
+        has_invited_or_paid_attendees = any(
+            (attendee.is_active and (attendee.invite or attendee.pay)) for attendee in attendees
+        )
+
+        if has_invited_or_paid_attendees:
+            raise BadRequestError("Cannot delete event with invited or paid attendees.")
+
+        try:
             for attendee in attendees:
                 self.attendee_service.delete_attendee(attendee.id)
